a-eye/requirements.txt
PyQt6>=6.6.0
mss>=9.0.0
Pillow>=10.0.0
anthropic>=0.25.0
openai>=1.30.0
google-generativeai>=0.5.0
keyring>=24.0.0
pyinstaller>=6.0.0
